+++ examples/v3.0/Cargo.toml
@@ -3,10 +3,6 @@ name = "examples_v3_0"
 version = "0.0.1"
 edition = "2021"
 
-[lib]
-name = "examples_v3_0"
-path = "./lib.rs"
-
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
