+++ examples/v3.0/src/components/mod.rs
@@ -1,0 +1,1 @@
+pub mod schemas;
