+++ examples/v3.0/src/components/schemas/all_of.yaml
