+++ examples/v3.0/src/components/schemas/array.yaml
