+++ examples/v3.0/src/components/schemas/boolean_field.yaml
