+++ examples/v3.0/src/components/schemas/camel_case_fields.rs
