+++ examples/v3.0/src/components/schemas/enums.rs
