+++ examples/v3.0/src/components/schemas/newtype_numeric.rs
