+++ examples/v3.0/src/components/schemas/nullable_field.rs
