+++ examples/v3.0/src/components/schemas/numeric_fields.rs
