+++ examples/v3.0/src/components/schemas/optional_field.rs
