+++ examples/v3.0/src/components/schemas/ref_items.rs
