+++ examples/v3.0/src/components/schemas/reserved_keywords.rs
