+++ examples/v3.0/src/components/schemas/struct_simple.yaml
