+++ examples/v3.0/src/components/schemas/title_description.yaml
