+++ examples/v3.0/src/lib.rs
