+++ examples/v3.0/tests/components/mod.rs
@@ -1,9 +1,4 @@
-mod camel_case_fields;
-mod enums;
-mod newtype;
-mod nullable_field;
-mod reserved_keywords;
-mod struct_simple;
+mod schemas;
 
 fn flatten(x: &str) -> String {
     x.replace(&[' ', '\n'], "")
