+++ examples/v3.0/tests/components/schemas/mod.rs
@@ -1,0 +1,6 @@
+mod camel_case_fields;
+mod enums;
+mod newtype;
+mod nullable_field;
+mod reserved_keywords;
+mod struct_simple;
